requirements.txt
django==3.2.5
pillow
django-crispy-forms
django-rest-framework